refactor: SearchVC에서 View 관련 내용들을 HomeView로 분리
