+++ iOS/Airbnb_iOS/Airbnb_iOS/ViewController/SearchViewController.swift
@@ -11,117 +11,15 @@ class SearchViewController: UIViewController {
     
     let searchBar = UISearchBar()
     let nextVC = BrowseViewController()
-
-    private lazy var collectionView: UICollectionView = {
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: self.getCollectionViewLayout())
-        collectionView.isScrollEnabled = true
-        collectionView.showsHorizontalScrollIndicator = false
-        collectionView.showsVerticalScrollIndicator = false
-        collectionView.contentInset = .zero
-        collectionView.backgroundColor = .white
-        collectionView.clipsToBounds = true
-
-        collectionView.register(MainEventCell.self, forCellWithReuseIdentifier: MainEventCell.identifier)
-        collectionView.register(AroundSpotCell.self, forCellWithReuseIdentifier: AroundSpotCell.identifier)
-        collectionView.register(ThemeSpotCell.self, forCellWithReuseIdentifier: ThemeSpotCell.identifier)
-
-        collectionView.register(CollectionHeaderView.self, forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: CollectionHeaderView.identifier)
-        collectionView.translatesAutoresizingMaskIntoConstraints = false
-        return collectionView
-    }()
-
+    private lazy var homeView = HomeView(frame: view.frame)
     private let dataSource = SearchViewCollectionDataSource()
 
     override func viewDidLoad() {
         super.viewDidLoad()
         self.view.backgroundColor = .white
         self.setSearchBar()
-        setCollectionViewLayout()
-        collectionView.dataSource = dataSource
-    }
-
-    private func setCollectionViewLayout() {
-        view.addSubview(collectionView)
-        NSLayoutConstraint.activate([
-            collectionView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            collectionView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            collectionView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
-            collectionView.topAnchor.constraint(equalTo: view.topAnchor)
-        ])
-    }
-
-    private func getCollectionViewLayout() -> UICollectionViewCompositionalLayout {
-        UICollectionViewCompositionalLayout { (section, env) -> NSCollectionLayoutSection? in
-            switch section {
-            case 0:
-                let itemFractionalWidthFraction = 1.0
-                let groupFractionalHeightFraction = 1.0 / 2.0
-                let itemInset: CGFloat = 0
-
-                let itemSize = NSCollectionLayoutSize(
-                    widthDimension: .fractionalWidth(itemFractionalWidthFraction),
-                    heightDimension: .fractionalHeight(1))
-                let item = NSCollectionLayoutItem(layoutSize: itemSize)
-                item.contentInsets = NSDirectionalEdgeInsets(top: itemInset, leading: itemInset, bottom: 40, trailing: itemInset)
-
-                let groupSize = NSCollectionLayoutSize(
-                    widthDimension: .fractionalWidth(1.0),
-                    heightDimension: .fractionalHeight(groupFractionalHeightFraction))
-                let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
-                group.edgeSpacing = NSCollectionLayoutEdgeSpacing(leading: .flexible(0), top: nil, trailing: nil, bottom: nil)
-
-                let section = NSCollectionLayoutSection(group: group)
-                section.contentInsets = NSDirectionalEdgeInsets(top: itemInset, leading: itemInset, bottom: itemInset, trailing: itemInset)
-
-                let headerSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0), heightDimension: .absolute(50.0))
-                let header = NSCollectionLayoutBoundarySupplementaryItem(layoutSize: headerSize, elementKind: UICollectionView.elementKindSectionHeader, alignment: .top)
-                section.boundarySupplementaryItems = [header]
-
-                return section
-
-            case 1:
-                let itemFractionalWidthFraction = 0.8
-                let itemInset: CGFloat = 2.5
-
-                let itemSize = NSCollectionLayoutSize(
-                    widthDimension: .fractionalWidth(itemFractionalWidthFraction),
-                    heightDimension: .fractionalHeight(1))
-                let item = NSCollectionLayoutItem(layoutSize: itemSize)
-                item.contentInsets = NSDirectionalEdgeInsets(top: itemInset, leading: 16, bottom: 24, trailing: itemInset)
-
-                let group = NSCollectionLayoutGroup.vertical(layoutSize: .init(widthDimension: .fractionalWidth(0.75), heightDimension: .estimated(200)), subitem: item, count: 2)
-
-                let section = NSCollectionLayoutSection(group: group)
-                section.contentInsets = NSDirectionalEdgeInsets(top: itemInset, leading: itemInset, bottom: itemInset, trailing: itemInset)
-                let headerSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0), heightDimension: .absolute(50.0))
-                let header = NSCollectionLayoutBoundarySupplementaryItem(layoutSize: headerSize, elementKind: UICollectionView.elementKindSectionHeader, alignment: .top)
-                section.boundarySupplementaryItems = [header]
-                section.orthogonalScrollingBehavior = .continuous
-
-                return section
-
-            default:
-                let itemFractionalWidthFraction = 0.8
-                let itemInset: CGFloat = 2.5
-
-                let itemSize = NSCollectionLayoutSize(
-                    widthDimension: .fractionalWidth(itemFractionalWidthFraction),
-                    heightDimension: .fractionalHeight(1))
-                let item = NSCollectionLayoutItem(layoutSize: itemSize)
-                item.contentInsets = NSDirectionalEdgeInsets(top: 28, leading: 16, bottom: 24, trailing: 0)
-
-                let group = NSCollectionLayoutGroup.vertical(layoutSize: .init(widthDimension: .estimated(335), heightDimension: .estimated(420)), subitem: item, count: 1)
-
-                let section = NSCollectionLayoutSection(group: group)
-                section.contentInsets = NSDirectionalEdgeInsets(top: itemInset, leading: itemInset, bottom: itemInset, trailing: itemInset)
-                let headerSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0), heightDimension: .absolute(50.0))
-                let header = NSCollectionLayoutBoundarySupplementaryItem(layoutSize: headerSize, elementKind: UICollectionView.elementKindSectionHeader, alignment: .top)
-                section.boundarySupplementaryItems = [header]
-                section.orthogonalScrollingBehavior = .continuous
-
-                return section
-            }
-        }       
+        view = homeView
+        homeView.setDataSource(dataSource)
     }
 }
